Stop consumers when master receives SIGTERM
Sneakers already does that, but there is a bug when more than one
workers are being used (and we always use multiple workers).

The bug causes all workers to consume an additional message, wait for a
minute and then exit. This is not what we want - we want for the workers
to exit as soon as they finish their current job.